add pending requests handling in example app
